[Refactor] payment 관련 dto 폴더 변경
