+++ backend/src/main/java/com/creationstack/backend/dto/Payment/DeletePaymentMethodRequestDto.java
@@ -1,4 +1,4 @@
-package com.creationstack.backend.dto;
+package com.creationstack.backend.dto.Payment;
 
 import lombok.AllArgsConstructor;
 import lombok.Getter;
@@ -7,6 +7,7 @@ import lombok.Setter;
 @Getter
 @Setter
 @AllArgsConstructor
-public class SavePaymentMethodRequestDto {
+public class DeletePaymentMethodRequestDto {
   private String billingKey;
+  private String reason;
 }
